program_runtime: split out loading task code into a separate module
